in X11: added some methods related to key handling of the X server and added the XTest library/extension
git-svn-id: https://svn.java.net/svn/jna~svn/trunk@741 2f8a963e-d2e4-e7d0-97bf-ccb7fcea9d80

+++ jnalib/src/com/sun/jna/examples/KeyboardUtils.java
@@ -153,7 +153,7 @@ public class KeyboardUtils {
                     int idx = code / 8;
                     int shift = code % 8;
                     if ((keys[idx] & (1 << shift)) != 0) {
-                        int sym = lib.XKeycodeToKeysym(dpy, (byte)code, 0);
+                        int sym = lib.XKeycodeToKeysym(dpy, (byte)code, 0).intValue();
                         if (sym == keysym)
                             return true;
                     }
